feat: refactor HTTP client and remove deprecated orval-http-client
